pyproject.toml
[tool.poetry]
name = "delibera-api"
version = "0.1.0"
description = "Backend API para o sistema de votacao de assembleias"
authors = ["Delibera Team"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
pydantic = { version = "^2.7.0", extras = ["email"] }
pydantic-settings = "^2.3.0"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
bcrypt = "^3.2.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
